Polyglot engine refactoring
